Visual Studio 2013 update
This updates the project files for Visual Studio 2013, and removes
the old Visual Studio 6 (1998) project files.  The update tool had
a number of complaints (see UpgradeLog.htm) that may need to be
addressed.

Also, replaced .cvsignore with .gitignore.

Visual Studio reports 1886 build errors, nearly all of them due to
the switch from MBCS to Unicode.  The former is no longer
supported "out of the box", and its use is discouraged, so we're
going to bite the bullet and use wide characters in the UI.