Fix subvol charset handling
We weren't doing a MOR-to-UNI conversion on the sub-volume name, so
HFS volumes with non-ASCII characters didn't look right.

This also relocates the character-conversion code to a new source
file.  It's currently part of the reformat lib, though it arguably
belongs in util (but that would introduce a new dependency
between reformat and util).