Tweak "open file" behavior
For a file with ".shk" extension, prioritize disk image over NuFX
file archive.  The disk image code won't accept it if it has more
than one entry, and it's pretty rare for somebody to want to open
a single-disk archive in file mode.

Ideally all single-disk archives are named ".sdk", but there's
probably a lot that are just named ".shk", so this will probably
be more convenient.