Handle carriage returns in REM statements
With a bit of hex-editing it's possible to embed carriage returns
in REM statements.  The reformatter wasn't handling that well.  The
new output matches what LIST generates.

The output generated cannot be imported by the text-to-BASIC
importer because it doesn't understand the blank lines.  The output
generated before this change didn't work either, though that was a
bit harder to figure out because the CRs are harder to see in Windows
than CRLF.

It should be possible to teach the importer to handle such files,
though I think these files are pretty rare -- I happened to find
them in some Peter Watson freeware.