Half-step toward VS2015 CE
When initially opened in Visual Studio 2015 Community Edition, the
project was updated to use the v140_xp toolset.  When the program
was run under WinXP it complained about a missing runtime DLL.  When
the DLL was provided, it complained about another one (with a
slightly strange name).  So I reverted the tools to v120_xp, i.e.
Visual Studio 2013.  (I don't know if this works because the tools
are included with VS2015, or because I have VS2013 installed and it
managed to find them.)

Whatever the case, it now builds for me with either IDE, and seems
to work fine on Windows XP, but I'd like to figure out why the XP
build isn't working with the v140_xp tools.