Change file extraction extension handling
When extracting files, you can ask CiderPress to add a file extension
by checking the "add file extension" box.  Whether or not this box
is checked, files that undergo a format conversion (e.g. AWP to RTF,
or SHR to BMP) have a file extension added to identify the file's
new format.

This turned out to be confusing and inconvenient at times, notably
when working with Merlin source files.  See issues #10 and #26 on
github for details.

CiderPress no longer adds file extensions to format-converted files
unless the "add file extension" box is checked.  (The "Configure for
easy access in Windows" button checks this box for you, so the
default behavior is still "safe".)

Also, fix a minor visual glitch in the extract dialog.

Also, update to version 4.0.3-a3.